LPS-108280 Move soy layout miller columns to soy folder
