+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/miller_columns/soy/Layout.es.js
@@ -20,7 +20,7 @@ import {Config} from 'metal-state';
 import './LayoutBreadcrumbs.es';
 
 import './LayoutColumn.es';
-import {setIn} from '../utils/utils.es';
+import {setIn} from '../../utils/utils.es';
 import templates from './Layout.soy';
 import {
 	DROP_TARGET_BORDERS,
