+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/miller_columns/soy/Layout.soy
@@ -16,7 +16,7 @@
 	{@param _handleLayoutColumnItemCheckboxClick: any}
 	{@param _handleLayoutColumnItemClick: any}
 
-	<div ref="layout">
+	<div class="layout-container" ref="layout">
 		{call com.liferay.layout.admin.web.LayoutBreadcrumbs.render}
 			{param breadcrumbEntries: $breadcrumbEntries /}
 		{/call}
