+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/miller_columns/soy/LayoutColumn.es.js
