+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/miller_columns/soy/utils/LayoutUpdateUtils.es.js
@@ -12,7 +12,7 @@
  * details.
  */
 
-import {setIn} from '../../utils/utils.es';
+import {setIn} from '../../../utils/utils.es';
 import {DROP_TARGET_ITEM_TYPES} from './LayoutDragDrop.es';
 import {
 	getColumnActiveItem,
