+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/view_layouts_soy.jsp
@@ -67,7 +67,7 @@ LayoutsAdminManagementToolbarDisplayContext layoutsManagementToolbarDisplayConte
 
 					<soy:component-renderer
 						context="<%= context %>"
-						module="js/miller_columns/Layout.es"
+						module="js/miller_columns/soy/Layout.es"
 						templateNamespace="com.liferay.layout.admin.web.Layout.render"
 					/>
 				</c:otherwise>
